+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.2 under development
 
-- Bug #59: Fix incorrect check error occurred while converting IP address to bits representation in PHP 8.0+ (@vjik)
+- Bug #59: Fix error while converting IP address to bits representation in PHP 8.0+ (@vjik)
 
 ## 1.0.1 January 27, 2022