+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## 1.0.2 under development
 
-- no changes in this release.
+- Bug #59: Fix error while converting IP address to bits representation in PHP 8.0+ (@vjik)
 
 ## 1.0.1 January 27, 2022
 
-- Bug #40: Fix return type for callback of `set_error_handler()` function (devanych)
+- Bug #40: Fix return type for callback of `set_error_handler()` function (@devanych)
 
 ## 1.0.0 March 04, 2021