+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ## 1.0.2 under development
 
 - New #63: Add `IpRanges` that represents a set of IP ranges that are either allowed or forbidden (@vjik)
+- Bug #59: Fix error while converting IP address to bits representation in PHP 8.0+ (@vjik)
 
 ## 1.0.1 January 27, 2022
 
-- Bug #40: Fix return type for callback of `set_error_handler()` function (devanych)
+- Bug #40: Fix return type for callback of `set_error_handler()` function (@devanych)
 
 ## 1.0.0 March 04, 2021