+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii Network Utilities Change Log
 
-## 1.0.2 under development
+## 1.1.0 August 06, 2024
 
 - New #63: Add `IpRanges` that represents a set of IP ranges that are either allowed or forbidden (@vjik)
 - Bug #59: Fix error while converting IP address to bits representation in PHP 8.0+ (@vjik)